smal refactoring
